requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

beautifulsoup4==4.13.4
dropbox==12.0.2
joblib==1.5.1
numpy==1.26.4
pandas==2.3.1
pydantic==2.11.7
pydantic-settings==2.10.1
pytest==8.4.1
python-dotenv==1.1.1
requests==2.32.4
scikit-learn==1.7.1
streamlit==1.47.1
xgboost==3.0.3
